pyproject.toml
[tool.poetry]
name = "birdcall_distribution"
version = "0.2.0"
description = ""
authors = ["Anthony Miyaguchi <acmiyaguchi@gatech.edu>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
matplotlib = "^3.5.3"
pandas = "^1.5.1"
pymc = "^4.3.0"
arviz = "*"
Cartopy = "^0.21.0"
earthengine-api = "^0.1.330"
tqdm = "^4.64.1"
pyarrow = "^10.0.0"
graphviz = "^0.20.1"
scikit-learn = "^1.1.3"
jax = {extras = ["gpu"], version = "^0.3.24"}
jaxlib = "^0.3.24+cuda11.cudnn82"
numpyro = "^0.10.1"


[tool.poetry.dev-dependencies]
jupyterlab = "^3.4.5"

[tool.poetry.group.dev.dependencies]
black = {version = "^22.10.0", allow-prereleases = true}


[[tool.poetry.source]]
name = "jax"
url = "https://storage.googleapis.com/jax-releases/jax_cuda_releases.html"
default = false
secondary = false

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
